pytorch_autoencoder/pyproject.toml
[tool.poetry]
name = "curi"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.5"
matplotlib = "^3.3.3"
scikit-learn = "^0.24.1"
torch = "^1.7.1"
pandas = "^1.2.1"
ipython = "^7.19.0"
jupyter = "^1.0.0"
jupyterlab = "^3.0.5"
torchvision = "^0.8.2"
icecream = "^2.0.0"
mlflow = "^1.13.1"
pytorch-lightning = "^1.1.5"
tensorboard = "^2.4.1"
Pillow = "^8.1.0"
scikit-image = "^0.18.1"
seaborn = "^0.11.1"
tqdm = "^4.56.0"
wfdb = "^3.2.0"
torchinfo = "^1.4.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
